+++ template-currey.php
@@ -7,7 +7,7 @@
 
            <div id="brand-page_heading_image" class="col-md-8">
 			  <div class="prod_cat_img">
-                <img style="height: auto; width: 100%;" src="">
+                <img style="height: auto; width: 100%;" src="<?php echo site_url('/wp-content/uploads/2020/08/6000-0192_8__brand_page.jpg'); ?>">
               </div>
 			</div>
 
@@ -15,7 +15,7 @@
             <div class="row h-100 justify-content-center">
                 <div class="col-lg-12 col-12 my-auto">
                     <div class="prod_cat_text" style="padding: 10px;">
-                        <h2 class="prod-cat-title">Currey & Co.</h2>
+                        <h2 class="prod-cat-title">Currey & Company</h2>
                     </div>
                 </div>
             </div>
@@ -42,7 +42,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/currey-co-mirrors'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1000-0076_2_-scaled.jpg">', wp_get_attachment_metadata(173289), 173289 ); ?>
 								 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Mirrors</h3>
                                 </div>
@@ -55,7 +55,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/currey-co-accessories/currey-co-sculptures-and-figurines'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(), ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0274_1_.jpg">', wp_get_attachment_metadata(172941), 172941); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Sculptures and Figurines</h3>
                                 </div>
@@ -68,7 +68,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/currey-co-accessories/currey-co-bowls-and-trays'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0251_1_-scaled.jpg">', wp_get_attachment_metadata(173270), 173270 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Decorative Bowls and Trays</h3>
                                 </div>
@@ -81,7 +81,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/currey-co-accessories/currey-co-boxes-and-containers'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0285_2_.jpg">', wp_get_attachment_metadata(173259), 173259 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Decorative Boxes and Containers</h3>
                                 </div>
@@ -94,7 +94,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/currey-co-accessories/currey-co-vases-and-bottles'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0352_2_.jpg">', wp_get_attachment_metadata(173200), 173200 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Decorative Vases and Bottles</h3>
                                 </div>
@@ -107,7 +107,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-home-decor/'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0262_3_-scaled.jpg">', wp_get_attachment_metadata(172937), 172937 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">All Decor</h3>
                                 </div>
@@ -120,7 +120,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co-home-decor/currey-co-accessories/'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(), ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/1200-0313_2_-scaled.jpg">', wp_get_attachment_metadata(173060), 173060 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">All Accessories</h3>
                                 </div>
@@ -133,7 +133,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-lamps/currey-co-table-lamps'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(), ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/6000-0631_2_-scaled.jpg">', wp_get_attachment_metadata(173518), 173518); ?>
 								 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Table Lamps</h3>
                                 </div>
@@ -146,7 +146,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-lamps/currey-co-floor-lamps'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="wp-content/uploads/2020/08/8000-0085_1_-scaled.jpg">', wp_get_attachment_metadata(173539), 173539); ?>
 								 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Floor Lamps</h3>
                                 </div>
@@ -159,7 +159,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-pendants'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/9000-0658_1_-scaled.jpg">', wp_get_attachment_metadata(173628), 173628); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Pendants</h3>
                                 </div>
@@ -172,7 +172,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-chandeliers'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/9000-0630_1_-scaled.jpg">', wp_get_attachment_metadata(173602), 173602); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Chandeliers</h3>
                                 </div>
@@ -185,7 +185,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-flush-mounts'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/9999-0051_1_.jpg">', wp_get_attachment_metadata(173641), 173641); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Flush Mounts</h3>
                                 </div>
@@ -198,7 +198,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting/currey-co-sconces'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/5000-0162_1_.jpg">', wp_get_attachment_metadata(173306), 173306); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">Sconces</h3>
                                 </div>
@@ -211,7 +211,7 @@
 
                         <a href="<?php echo site_url('/product-category/currey-co/currey-co-lighting'); ?>">
                             <div class="brand-page_cat_image_container pt-3">
-                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="">', wp_get_attachment_metadata(),  ); ?>
+                                <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/08/9000-0652_1_-scaled.jpg">', wp_get_attachment_metadata(173623), 173623 ); ?>
                                 <div class="pb-3 pt-3">
                                     <h3 class="brand-page_cat_headings" style="text-align: center;" class="mb-2">All Lighting</h3>
                                 </div>